requirements.txt
bcrypt
black
fastapi
passlib
pre-commit
pydantic-settingspylint
python-jose
python-multipart
sqlalchemy
uvicorn
